Split `SnowFlakeBenchmark` into separate benchmarks.
Rename `SnowFlakeBenchmark` to `AtomicSnowFlakeBenchmark`.
Add `SnowFlakeBenchmark` for `SnowFlake`.
